stitch: Rewrite language on top of Javascript
Javascript is easier to read, and more familiar to users than lisp.
Moving to Javascript also removes the development overhead of writing
language features not related to our project, such as functions and
mathematical operators.

The language is now represented as Javascript objects (e.g. Machine,
Docker, Connection). Various functions create these objects and save
them to a context variable, which is later queried by the Quilt engine
to enforce specs as before.