provider: Fix Amazon ACL generation (Squash before merge)
Before, if a cluster with all AmazonReserved or AmazonSpot was booted,
the security groups to the cluster would fluctuate because they shared
the same ACL-setting logic. This patch makes it so if either one of
those providers have machines in the cluster, the other will also set
the desired ACls.